modularized account settings page
